A BMI calculator that logs results to the console
A BMI (Body Mass Index) Calculator that logs its results to the console. 
This is a way to determine a generalized estimate of an adult's body mass (fat, bone, and muscle) in relation to their height.
The calculator takes weight in pounds, and height in inches; converting these into: kilograms and meters, respectively.
Once the BMI is calculated, the results are assigned one of the following categories: underweight, normal weight, overweight , or obese.